refactor(shell): split each shell into its own module directory
