--- crates/atuin-common/src/shell/fish.rs
@@ -1,290 +1,0 @@
-use std::{
-    collections::HashMap,
-    path::{Path, PathBuf},
-    process,
-    sync::Arc,
-};
-
-use futures::{
-    FutureExt,
-    future::{BoxFuture, Shared},
-};
-use tracing::instrument;
-use winnow::{
-    ModalResult, Parser,
-    combinator::{alt, delimited, opt, preceded, repeat, terminated},
-    token::{literal, take_while},
-};
-
-use bstr::BString;
-
-use super::{AliasValue, AliasesError, IsShell, RunError};
-
-type Aliases = HashMap<BString, AliasValue>;
-
-type Probe<T, E> = Shared<BoxFuture<'static, Result<T, E>>>;
-
-/// The `fish` executable itself, and the ability to invoke it.
-#[derive(Debug)]
-struct FishExe {
-    path: PathBuf,
-}
-
-impl FishExe {
-    /// `fish -i` sources the user's config files before running our command, and anything they
-    /// print lands on the same stdout. Bracket the real output with these NUL-delimited markers so
-    /// it can be sliced back out; NUL cannot occur in a command's arguments or output.
-    const OUTPUT_BEGIN: &[u8] = b"\0atuin\0";
-    const OUTPUT_END: &[u8] = b"\0nituA\0";
-
-    fn new(path: impl Into<PathBuf>) -> Self {
-        Self { path: path.into() }
-    }
-
-    fn path(&self) -> &Path {
-        &self.path
-    }
-
-    /// Wrap `command` so its output is delimited by [`Self::OUTPUT_BEGIN`] and
-    /// [`Self::OUTPUT_END`]. `$status` is captured and re-raised so that framing does not mask the
-    /// command's own exit status.
-    fn frame(command: &str) -> String {
-        format!(
-            r"printf '\000atuin\000'; {command}; set __atuin_status $status; printf '\000nituA\000'; exit $__atuin_status"
-        )
-    }
-
-    #[instrument(skip(command))]
-    async fn run(&self, command: &str) -> Result<process::Output, RunError> {
-        let mut output = tokio::process::Command::new(&self.path)
-            .args(["-ic", &Self::frame(command)])
-            .output()
-            .await
-            .map_err(|error| RunError::Io {
-                command: command.to_owned(),
-                error: Arc::new(error),
-            })?;
-
-        let delimiter = || RunError::Delimiter {
-            command: command.to_owned(),
-        };
-        let start = output
-            .stdout
-            .windows(Self::OUTPUT_BEGIN.len())
-            .position(|window| window == Self::OUTPUT_BEGIN)
-            .map(|at| at + Self::OUTPUT_BEGIN.len())
-            .ok_or_else(delimiter)?;
-        let end = output.stdout[start..]
-            .windows(Self::OUTPUT_END.len())
-            .position(|window| window == Self::OUTPUT_END)
-            .map(|at| at + start)
-            .ok_or_else(delimiter)?;
-
-        output.stdout = output.stdout[start..end].to_vec();
-
-        if output.status.success() {
-            Ok(output)
-        } else {
-            Err(RunError::Exec {
-                command: command.to_owned(),
-                status: output.status,
-                stdout: output.stdout.into(),
-                stderr: output.stderr.into(),
-            })
-        }
-    }
-}
-
-#[derive(Debug)]
-struct Inner {
-    config_path: PathBuf,
-    aliases: Probe<Aliases, AliasesError>,
-}
-
-#[derive(Debug, Clone, derive_more::Display)]
-#[display("fish")]
-pub struct Fish {
-    exe: Arc<FishExe>,
-    inner: Arc<Inner>,
-}
-
-impl Fish {
-    /// Create a new Fish shell object.
-    ///
-    /// This will kick off background tokio tasks to eagerly probe information. Resolving the
-    /// asynchronous methods will block on said probe tasks.
-    pub fn new(path: &Path) -> Self {
-        let config_path = directories::BaseDirs::new()
-            .map(|dirs| dirs.home_dir().join(".config/fish/config.fish"))
-            .unwrap_or_else(|| PathBuf::from(".config/fish/config.fish"));
-
-        let exe = Arc::new(FishExe::new(path));
-
-        let probe_exe = exe.clone();
-        let probe = tokio::spawn(async move {
-            let output = probe_exe.run("alias").await?;
-            Self::parse_aliases(&output.stdout)
-        });
-
-        let aliases = async move { probe.await.unwrap_or(Err(AliasesError::Probe)) }
-            .boxed()
-            .shared();
-
-        Self {
-            exe,
-            inner: Arc::new(Inner {
-                config_path,
-                aliases,
-            }),
-        }
-    }
-
-    /// Parse the output of `alias` into a name → value map.
-    fn parse_aliases(input: &[u8]) -> Result<Aliases, AliasesError> {
-        fn piece<'a>(input: &mut &'a [u8]) -> ModalResult<&'a [u8]> {
-            alt((
-                preceded(
-                    b'\\',
-                    alt((literal(br"\".as_slice()), literal(b"'".as_slice()))),
-                ),
-                take_while(1.., |b: u8| b != b'\'' && b != b'\\'),
-                literal(br"\".as_slice()),
-            ))
-            .parse_next(input)
-        }
-
-        fn quoted_value(input: &mut &[u8]) -> ModalResult<Vec<u8>> {
-            delimited(
-                b'\'',
-                repeat(0.., piece).fold(Vec::new, |mut acc: Vec<u8>, seg: &[u8]| {
-                    acc.extend_from_slice(seg);
-                    acc
-                }),
-                b'\'',
-            )
-            .parse_next(input)
-        }
-
-        fn bare_value(input: &mut &[u8]) -> ModalResult<Vec<u8>> {
-            take_while(1.., |b: u8| b != b' ' && b != b'\n')
-                .map(<[u8]>::to_vec)
-                .parse_next(input)
-        }
-
-        fn alias_line(input: &mut &[u8]) -> ModalResult<(Vec<u8>, Vec<u8>)> {
-            (
-                literal(b"alias ".as_slice()),
-                take_while(1.., |b: u8| b != b' ' && b != b'\n'),
-                literal(b" ".as_slice()),
-                alt((quoted_value, bare_value)),
-            )
-                .map(|(_, name, _, value): (_, &[u8], _, Vec<u8>)| (name.to_vec(), value))
-                .parse_next(input)
-        }
-
-        let mut records = repeat(0.., terminated(alias_line, opt(literal(b"\n".as_slice())))).fold(
-            HashMap::new,
-            |mut acc: Aliases, (name, value)| {
-                acc.insert(
-                    BString::from(name),
-                    AliasValue::Command(BString::from(value)),
-                );
-                acc
-            },
-        );
-
-        records.parse(input).map_err(|error| {
-            let offset = error.offset();
-            let near = input[offset..].iter().take(48).copied().collect::<Vec<_>>();
-
-            AliasesError::Parse {
-                offset,
-                near: String::from_utf8_lossy(&near).into_owned(),
-            }
-        })
-    }
-}
-
-#[async_trait::async_trait]
-impl IsShell for Fish {
-    fn canonical_name(&self) -> &'static str {
-        "fish"
-    }
-
-    fn is_posix(&self) -> bool {
-        false
-    }
-
-    #[instrument]
-    async fn aliases(&self) -> Result<HashMap<BString, AliasValue>, AliasesError> {
-        self.inner.aliases.clone().await
-    }
-
-    #[instrument(skip(command))]
-    async fn run_interactive(&self, command: &str) -> Result<process::Output, RunError> {
-        self.exe.run(command).await
-    }
-
-    fn installed_path(&self) -> Option<&Path> {
-        Some(self.exe.path())
-    }
-
-    fn user_config_path(&self) -> &Path {
-        &self.inner.config_path
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-
-    fn parse(input: &[u8]) -> HashMap<BString, BString> {
-        Fish::parse_aliases(input)
-            .unwrap()
-            .into_iter()
-            .map(|(k, v)| match v {
-                AliasValue::Command(cmd) => (k, cmd),
-                AliasValue::Argv(_) => unreachable!("fish yields command strings"),
-            })
-            .collect()
-    }
-
-    #[test]
-    fn parses_bare_and_quoted() {
-        assert_eq!(
-            parse(b"alias plain man\n")[&BString::from(&b"plain"[..])],
-            BString::from(&b"man"[..])
-        );
-        assert_eq!(
-            parse(b"alias ll 'ls -l'\n")[&BString::from(&b"ll"[..])],
-            BString::from(&b"ls -l"[..])
-        );
-    }
-
-    #[test]
-    fn decodes_backslash_escapes() {
-        assert_eq!(
-            parse(br"alias q 'it\'s'")[&BString::from(&b"q"[..])],
-            BString::from(&b"it's"[..])
-        );
-        assert_eq!(
-            parse(br"alias b 'a\\b'")[&BString::from(&b"b"[..])],
-            BString::from(&br"a\b"[..])
-        );
-    }
-
-    #[test]
-    fn does_not_use_equals_as_separator() {
-        // fish records are `alias NAME VALUE`; an `=` is just an ordinary byte.
-        assert_eq!(
-            parse(b"alias k 'a=b'\n")[&BString::from(&b"k"[..])],
-            BString::from(&b"a=b"[..])
-        );
-    }
-
-    #[test]
-    fn rejects_trailing_garbage() {
-        assert!(Fish::parse_aliases(b"alias ll 'ls -l'\nnonsense\n").is_err());
-    }
-}
